requirements.txt
aiohttp~=3.9.0
pyrofork==2.3.50
beautifulsoup4~=4.8.0
requests
pypdf
Pillow~=10.1.0
tgcrypto
sqlmodel~=0.0.6
asyncpg
psycopg2-binary
aiosqlite
reportlab
SQLAlchemy
psutil~=5.9.0
alembic
loguru
flask
gunicorn
